wip - add regression tests
