+++ tests/Integration/View/Fixtures/stacktrace-component-imported-usage.view.php
@@ -1,0 +1,5 @@
+<?php
+use function Tempest\View\view;
+?>
+
+<x-stacktrace-error-component />
