+++ tests/Integration/View/Fixtures/stacktrace-standalone-error.view.php
@@ -1,0 +1,3 @@
+<div>
+    <?php missing_inline_view_function(); ?>
+</div>
