+++ tests/Integration/View/Fixtures/x-stacktrace-error-component.view.php
@@ -1,0 +1,3 @@
+<section>
+    <?php missing_component_stacktrace_function(); ?>
+</section>
