+++ tests/Integration/View/ViewComponentTest.php
@@ -8,7 +8,11 @@ use Generator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use Tempest\Core\Environment;
+use Tempest\Http\GenericRequest;
+use Tempest\Http\Method;
 use Tempest\Http\Session\FormSession;
+use Tempest\Router\Exceptions\DevelopmentException;
+use Tempest\Router\Exceptions\HtmlExceptionRenderer;
 use Tempest\Validation\FailingRule;
 use Tempest\Validation\Rules\IsAlphaNumeric;
 use Tempest\Validation\Rules\IsBetween;
@@ -16,7 +20,10 @@ use Tempest\Validation\Validator;
 use Tempest\View\Exceptions\DataAttributeWasInvalid;
 use Tempest\View\Exceptions\ViewCompilationFailed;
 use Tempest\View\Exceptions\ViewVariableWasReserved;
+use Tempest\View\GenericView;
+use Tempest\View\Renderers\TempestViewRenderer;
 use Tempest\View\ViewCache;
+use Tempest\View\ViewConfig;
 use Tests\Tempest\Fixtures\Views\Chapter;
 use Tests\Tempest\Fixtures\Views\DocsView;
 use Tests\Tempest\Integration\FrameworkIntegrationTestCase;
@@ -1082,6 +1089,83 @@ final class ViewComponentTest extends FrameworkIntegrationTestCase
             function (ViewCompilationFailed $exception): void {
                 $this->assertStringContainsString('missing-import-view.view.php', $exception->getFile());
                 $this->assertSame(2, $exception->getLine());
+            },
+        );
+    }
+
+    #[Test]
+    public function does_not_duplicate_view_compilation_frames_in_stacktrace(): void
+    {
+        $this->container->singleton(Environment::class, Environment::LOCAL);
+        $this->container->singleton(GenericRequest::class, new GenericRequest(
+            method: Method::GET,
+            uri: '/',
+        ));
+
+        $viewRenderer = TempestViewRenderer::make();
+        $exceptionRenderer = $this->container->get(HtmlExceptionRenderer::class);
+
+        $this->assertException(
+            ViewCompilationFailed::class,
+            function () use ($viewRenderer): void {
+                $viewRenderer->render(view(__DIR__ . '/Fixtures/stacktrace-standalone-error.view.php'));
+            },
+            function (ViewCompilationFailed $exception) use ($exceptionRenderer): void {
+                $response = $exceptionRenderer->render($exception);
+
+                $this->assertInstanceOf(DevelopmentException::class, $response);
+                $this->assertInstanceOf(GenericView::class, $response->body);
+
+                $hydration = json_decode($response->body->data['hydration'], associative: true, flags: JSON_THROW_ON_ERROR);
+                $stacktrace = json_decode($hydration['stacktrace'], associative: true, flags: JSON_THROW_ON_ERROR);
+
+                $renderCompiledFrames = array_values(array_filter(
+                    $stacktrace['applicationFrames'],
+                    fn (array $frame): bool => ($frame['class'] ?? null) === TempestViewRenderer::class && ($frame['function'] ?? null) === 'renderCompiled',
+                ));
+
+                $this->assertCount(1, $renderCompiledFrames);
+            },
+        );
+    }
+
+    #[Test]
+    public function maps_component_source_line_in_view_compilation_stacktrace(): void
+    {
+        $this->container->singleton(Environment::class, Environment::LOCAL);
+        $this->container->singleton(GenericRequest::class, new GenericRequest(
+            method: Method::GET,
+            uri: '/',
+        ));
+
+        $this->container->get(ViewConfig::class)->addViewComponents(
+            __DIR__ . '/Fixtures/x-stacktrace-error-component.view.php',
+        );
+
+        $exceptionRenderer = $this->container->get(HtmlExceptionRenderer::class);
+
+        $this->assertException(
+            ViewCompilationFailed::class,
+            function (): void {
+                $this->view->render(view(__DIR__ . '/Fixtures/stacktrace-component-imported-usage.view.php'));
+            },
+            function (ViewCompilationFailed $exception) use ($exceptionRenderer): void {
+                $response = $exceptionRenderer->render($exception);
+
+                $this->assertInstanceOf(DevelopmentException::class, $response);
+                $this->assertInstanceOf(GenericView::class, $response->body);
+
+                $hydration = json_decode($response->body->data['hydration'], associative: true, flags: JSON_THROW_ON_ERROR);
+                $stacktrace = json_decode($hydration['stacktrace'], associative: true, flags: JSON_THROW_ON_ERROR);
+
+                $renderCompiledFrames = array_values(array_filter(
+                    $stacktrace['applicationFrames'],
+                    fn (array $frame): bool => ($frame['class'] ?? null) === TempestViewRenderer::class && ($frame['function'] ?? null) === 'renderCompiled',
+                ));
+
+                $this->assertCount(1, $renderCompiledFrames);
+                $this->assertSame('tests/Integration/View/Fixtures/x-stacktrace-error-component.view.php', $renderCompiledFrames[0]['relativeFile']);
+                $this->assertSame(2, $renderCompiledFrames[0]['line']);
             },
         );
     }
